+++ system/src/sheets/FalloutPcSheet.mjs
@@ -28,7 +28,6 @@ export default class FalloutPcSheet extends FalloutBaseActorSheet {
 			"disease",
 			"miscellany",
 			"perk",
-			"robot_mod",
 			"skill",
 			"trait",
 			"weapon_mod",
@@ -402,7 +401,7 @@ export default class FalloutPcSheet extends FalloutBaseActorSheet {
 
 		context.itemsEnrichedDescriptions = {};
 
-		for await (let item of context.itemsByType.robot_mod) {
+		for (let item of context.itemsByType.robot_mod ?? []) {
 			const descriptionRich = await TextEditor.enrichHTML(
 				item.system.effect, {async: true}
 			);